Fix a race in SocketContext.
The design was to enqueue the messages, then schedule the OutboundProxy write task.
In the event of high write volume, messages can be flushed faster, if IO-write thread is running.
However, we will have a race with SocketContext.destroy(). I think this race is not preventable,
unless we opt to another design: pass the message along with the scheduled write.
I dont want that, thus, fix the race by catching NullPointerException.